refactor: moved snipe and esnipe out of nypsi file
