+++ src/events/messageDelete.ts
@@ -1,6 +1,5 @@
 import { Message } from "discord.js";
-import { snipe } from "../nypsi";
-import { createGuild, getChatFilter, getSnipeFilter, hasGuild } from "../utils/guilds/utils";
+import { createGuild, getChatFilter, getSnipeFilter, hasGuild, snipe } from "../utils/guilds/utils";
 
 export default async function messageDelete(message: Message) {
     if (!message) return;
